requirements.txt
RPi.GPIO==0.6.3
pygame==1.9.4
rx
